requirements.txt
Flask
psycopg2-binary
Flask-Login
Flask-Bcrypt
python-dotenv
flask-sqlalchemy